minor fixes in the mvc refactor progress
